fixup! gh-137314: Fix incorrect treatment of format specs in raw fstrings (GH-137328) (cherry picked from commit 0153d82a5ab0c6ac16c046bdd4438ea11b58d59d)
